update: merge app.js into index.js
